refactor(src): move extension runtime helpers
