+++ src/extension-runtime/bundled-extension-paths.ts
