+++ src/extension-runtime/bundled-resource-path.ts
