+++ src/extension-runtime/extension-validator.ts
