+++ src/tests/extension-discovery.test.ts
@@ -6,7 +6,7 @@ import assert from 'node:assert/strict'
 import { mkdirSync, writeFileSync, rmSync } from 'node:fs'
 import { join } from 'node:path'
 import { tmpdir } from 'node:os'
-import { resolveExtensionEntries, discoverExtensionEntryPaths, mergeExtensionEntryPaths } from '../extension-discovery.ts'
+import { resolveExtensionEntries, discoverExtensionEntryPaths, mergeExtensionEntryPaths } from '../extension-runtime/extension-discovery.ts'
 
 function makeTempDir(): string {
   const dir = join(tmpdir(), `ext-discovery-test-${Date.now()}-${Math.random().toString(36).slice(2)}`)
